separate EntityResponse logic
